+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [7.0.11] - 2025-04-14
+### Changed
+- Added download survey results descriptions [#4950](https://github.com/rokwire/illinois-app/issues/4950).
+- Do not show "Take Attendance" link for events that have no attedance options set [#4994](https://github.com/rokwire/illinois-app/issues/4994).
+
 ## [7.0.10] - 2025-04-11
 ### Fixed
 - Fixed event detail panel update after successful self-checkin [#4956](https://github.com/rokwire/illinois-app/issues/4956).